% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_scoring.R
\name{condition_number}
\alias{condition_number}
\title{compute condition number of matrix}
\usage{
condition_number(info)
}
\arguments{
\item{info}{matrix}
}
\description{
compute condition number of matrix
}
