% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{expit}
\alias{expit}
\alias{intexpit}
\title{expit function and integral of expit function}
\usage{
expit(x)

intexpit(x)
}
\arguments{
\item{x}{argument to expit or intexpit function}
}
\description{
expit function and integral of expit function
}
