% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exponential_anisotropic3D_alt}
\alias{exponential_anisotropic3D_alt}
\alias{d_exponential_anisotropic3D_alt}
\title{Geometrically anisotropic exponential covariance function (three dimensions, alternate parameterization)}
\usage{
exponential_anisotropic3D_alt(covparms, locs)

d_exponential_anisotropic3D_alt(covparms, locs)
}
\arguments{
\item{covparms}{A vector with covariance parameters
in the form (variance, B11, B12, B13, B22, B23, B33, smoothness, nugget)}

\item{locs}{A matrix with \code{n} rows and \code{3} columns.
Each row of locs is a point in R^3.}
}
\value{
A matrix with \code{n} rows and \code{n} columns, with the i,j entry
containing the covariance between observations at \code{locs[i,]} and
\code{locs[j,]}.
}
\description{
From a matrix of locations and covariance parameters of the form
(variance, B11, B12, B13, B22, B23, B33, smoothness, nugget), return the square matrix of
all pairwise covariances.
}
\section{Functions}{
\itemize{
\item \code{d_exponential_anisotropic3D_alt}: Derivatives of anisotropic Matern covariance
}}

\section{}{
NA
}

