% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{fast_Gp_sim}
\alias{fast_Gp_sim}
\title{Approximate GP simulation}
\usage{
fast_Gp_sim(covparms, covfun_name = "matern_isotropic", locs, m = 30)
}
\arguments{
\item{covparms}{Vector of covariance function parameters. For "maternIsotropic" and
"maternSphere", these are (variance, range, smoothness, nugget). For "maternSphereTime",
these are (variance, spatial range, temporal range, smoothness, nugget). See documentation
of the covariance functions for more details.}

\item{covfun_name}{string name of covariance functions. Currently supported
covariance functions are "maternIsotropic", "maternSphere", and "maternSphereTime"}

\item{locs}{Matrix of locations. Each row is an individual location. For "maternSphere",
locations are (lon,lat), where lon is in (-180,180), and lat is in (-90,90). For "maternSphereTime"
locations are (lon,lat,time).}

\item{m}{Number of nearest neighbors to use in approximation}
}
\value{
vector of simulated values
}
\description{
Calculates an approximation to the inverse Cholesky
factor of the covariance matrix using Vecchia's approximation,
then the simulation is produced by solving a linear system
with a vector of uncorrelated standard normals
}
\examples{
locs <- as.matrix( expand.grid( (1:100)/100, (1:100)/100 ) )
y <- fast_Gp_sim(c(4,0.2,0.5,0), "matern_isotropic",  locs, 30 )
fields::image.plot( matrix(y,100,100) )
}
