% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbor_functions.R
\name{find_ordered_nn_brute}
\alias{find_ordered_nn_brute}
\title{Naive brute force nearest neighbor finder}
\usage{
find_ordered_nn_brute(locs, m)
}
\arguments{
\item{locs}{matrix of locations}

\item{m}{number of neighbors}
}
\value{
integer vector giving ordering
}
\description{
Naive brute force nearest neighbor finder
}
