% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_scoring.R
\name{fisher_scoring}
\alias{fisher_scoring}
\title{Fisher scoring algorithm}
\usage{
fisher_scoring(likfun, start_parms, link, silent = FALSE,
  convtol = 1e-04)
}
\arguments{
\item{likfun}{likelihood function, returns likelihood, gradient, and hessian}

\item{start_parms}{starting values of parameters}

\item{link}{link function for parameters (used for printing)}

\item{silent}{TRUE/FALSE for suppressing output}

\item{convtol}{convergence tolerance on step dot grad}
}
\description{
Fisher scoring algorithm
}
