% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{get_linkfun}
\alias{get_linkfun}
\title{get link function, whether locations are lonlat and space time}
\usage{
get_linkfun(covfun_name)
}
\arguments{
\item{covfun_name}{string name of covariance function}
}
\description{
get link function, whether locations are lonlat and space time
}
