% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{get_penalty}
\alias{get_penalty}
\title{get penalty function}
\usage{
get_penalty(y, X, locs, covfun_name)
}
\arguments{
\item{y}{response}

\item{X}{design matrix}

\item{locs}{locations}

\item{covfun_name}{string name of covariance function}
}
\description{
get penalty function
}
