% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{get_start_parms}
\alias{get_start_parms}
\title{get default starting values of covariance parameters}
\usage{
get_start_parms(y, X, locs, covfun_name)
}
\arguments{
\item{y}{response}

\item{X}{design matrix}

\item{locs}{locations}

\item{covfun_name}{string name of covariance function}
}
\description{
get default starting values of covariance parameters
}
