% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matern25_scaledim_relevance}
\alias{matern25_scaledim_relevance}
\title{Matern covariance function, smoothess = 2.5, different relevance parameter for each dimension}
\usage{
matern25_scaledim_relevance(covparms, locs)
}
\arguments{
\item{covparms}{A vector with covariance parameters
in the form (variance, relevance_1, ..., relevance_d, nugget)}

\item{locs}{A matrix with \code{n} rows and \code{d} columns.
Each row of locs is a point in R^d.}
}
\value{
A matrix with \code{n} rows and \code{n} columns, with the i,j entry
containing the covariance between observations at \code{locs[i,]} and
\code{locs[j,]}.
}
\description{
From a matrix of locations and covariance parameters of the form
(variance, relevance_1, ..., relevance_d, nugget), return the square matrix of
all pairwise covariances.
}
\section{Parameterization}{

The covariance parameter vector is (variance, relevance_1, ..., relevance_d, nugget).
The covariance function is parameterized as
\deqn{ M(x,y) = \sigma^2 (1 + || D(x - y) || + || D(x - y) ||^2/3.0) exp( - || D(x - y) || ) }
where D is a diagonal matrix with (relevance_1, ..., relevance_d) on the diagonals.
The nugget value \eqn{ \sigma^2 \tau^2 } is added to the diagonal of the covariance matrix.
NOTE: the nugget is \eqn{ \sigma^2 \tau^2 }, not \eqn{ \tau^2 }.
}

