% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering_functions.R
\name{order_coordinate}
\alias{order_coordinate}
\title{Sorted coordinate ordering}
\usage{
order_coordinate(locs, coordinate)
}
\arguments{
\item{locs}{A matrix of locations in R^d. Each row of \code{locs} contains a location.}
}
\value{
A vector of indices giving the ordering,
}
\description{
Return the ordering of locations sorted along one of the
coordinates or the sum of multiple coordinates
}
\examples{
n <- 100             # Number of locations
d <- 2               # dimension of domain
locs <- matrix( runif(n*d), n, d )
ord1 <- order_coordinate(locs, 1 )
ord12 <- order_coordinate(locs, c(1,2) )
}
