% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering_functions.R
\name{order_middleout}
\alias{order_middleout}
\title{Middle-out ordering}
\usage{
order_middleout(locs)
}
\arguments{
\item{locs}{A matrix of locations in R^d. Each row of \code{locs} contains a location.}
}
\value{
A vector of indices giving the ordering, i.e. 
the first element of this vector is the index of the location nearest the center.
}
\description{
Return the ordering of locations increasing in their
distance to the average location
}
\examples{
n <- 100             # Number of locations
d <- 2               # dimension of domain
locs <- matrix( runif(n*d), n, d )
ord <- order_middleout(locs)
}
