% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{pen_hi}
\alias{pen_hi}
\alias{dpen_hi}
\alias{ddpen_hi}
\title{penalize large values of parameter: penalty, 1st deriative, 2nd derivative}
\usage{
pen_hi(x, tt, aa)

dpen_hi(x, tt, aa)

ddpen_hi(x, tt, aa)
}
\arguments{
\item{x}{argument to penalty}

\item{tt}{scale parameter of penalty}

\item{aa}{location parameter of penalty}
}
\description{
penalize large values of parameter: penalty, 1st deriative, 2nd derivative
}
