% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{pen_lo}
\alias{pen_lo}
\alias{dpen_lo}
\alias{ddpen_lo}
\title{penalize small values of parameter: penalty, 1st deriative, 2nd derivative}
\usage{
pen_lo(x, tt, aa)

dpen_lo(x, tt, aa)

ddpen_lo(x, tt, aa)
}
\arguments{
\item{x}{argument to penalty}

\item{tt}{scale parameter of penalty}

\item{aa}{location parameter of penalty}
}
\description{
penalize small values of parameter: penalty, 1st deriative, 2nd derivative
}
