% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{pen_loglo}
\alias{pen_loglo}
\alias{dpen_loglo}
\alias{ddpen_loglo}
\title{penalize small values of log parameter: penalty, 1st deriative, 2nd derivative}
\usage{
pen_loglo(x, tt, aa)

dpen_loglo(x, tt, aa)

ddpen_loglo(x, tt, aa)
}
\arguments{
\item{x}{argument to penalty}

\item{tt}{scale parameter of penalty}

\item{aa}{location parameter of penalty}
}
\description{
penalize small values of log parameter: penalty, 1st deriative, 2nd derivative
}
