% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictions}
\alias{predictions}
\title{Compute Gaussian process predictions using Vecchia's approximations}
\usage{
predictions(covparms, covfun_name = "matern_isotropic", locs_obs, locs_pred,
  X_obs, X_pred, beta, y_obs, m = 30, lonlat = FALSE)
}
\arguments{
\item{covparms}{Covariance parameters}

\item{covfun_name}{Name of covariance function}

\item{locs_obs}{observation locations}

\item{locs_pred}{prediction locations}

\item{X_obs}{Design matrix for observations}

\item{X_pred}{Design matrix for predictions}

\item{beta}{Linear mean parameters}

\item{y_obs}{Observations assocaited with locs_obs}

\item{m}{Number of nearest neighbors to use}

\item{lonlat}{Flag indicating that locs contain longitudes and latitudes}
}
\description{
The prediction locations after the observation locations,
an approximation for the inverse Cholesky of the covariance matrix
is computed, and standard formulas are applied to obtain
the conditional expectation.
}
