% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sph_grad_xyz}
\alias{sph_grad_xyz}
\title{compute gradient of spherical harmonics functions}
\usage{
sph_grad_xyz(xyz, Lmax)
}
\arguments{
\item{xyz}{xyz coordinates of locations on sphere}

\item{Lmax}{largest degree of spherical harmonics. 
Current only Lmax=2 supported}
}
\description{
compute gradient of spherical harmonics functions
}
