% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{summary.GpGp_fit}
\alias{summary.GpGp_fit}
\title{Print summary of GpGp fit}
\usage{
\method{summary}{GpGp_fit}(fit)
}
\arguments{
\item{fit}{Object of class "GpGp_fit", usually the return value from 
\code{\link{fit_model}}}
}
\description{
Print summary of GpGp fit
}
