% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_scoring.R
\name{test_likelihood_object}
\alias{test_likelihood_object}
\title{test likelihood object for NA or Inf values}
\usage{
test_likelihood_object(likobj)
}
\arguments{
\item{likobj}{likelihood object}
}
\description{
test likelihood object for NA or Inf values
}
