% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering_functions.R
\name{order_maxmin}
\alias{order_maxmin}
\title{Maximum minimum distance ordering}
\usage{
order_maxmin(locs)
}
\arguments{
\item{locs}{A matrix of locations in R^d. Each row of \code{locs} contains a location.}
}
\value{
A vector of indices giving the ordering,
}
\description{
Return the indices of an approximation to the maximum minimum distance ordering.
A point in the center is chosen first, and then each successive point
is chosen to maximize the minimum distance to previously selected points
}
\examples{
nvec <- c(50,50)
locs <- as.matrix( expand.grid( 1:nvec[1]/nvec[1], 1:nvec[2]/nvec[2] ) )
ord <- order_maxmin(locs)
par(mfrow=c(1,3))
plot( locs[ord[1:100],1], locs[ord[1:100],2], xlim = c(0,1), ylim = c(0,1) )
plot( locs[ord[1:300],1], locs[ord[1:300],2], xlim = c(0,1), ylim = c(0,1) )
plot( locs[ord[1:900],1], locs[ord[1:900],2], xlim = c(0,1), ylim = c(0,1) )

}
